fix: Check if dir exists before deleting

+++ build.sh
@@ -1,5 +1,5 @@
 source bootstrap.sh
 source venv/bin/activate
-rm -rf _build
+[ -d _build ] && rm -rf _build
 make html
 deactivate
